requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.1
python-dotenv>=1.0
Pillow>=10.0
djangorestframework>=3.15
django-filter>=24.2
django-cors-headers>=4.4
drf-spectacular>=0.27
channels>=4.1
channels-redis>=4.2
celery[redis]>=5.4
redis>=5.0
python-decouple>=3.8
gunicorn>=22.0
uvicorn[standard]>=0.30
whitenoise>=6.7
daphne>=4.1

# Testing
pytest>=8.0
pytest-django>=4.8
pytest-asyncio>=0.23
